receiver8/requirements.txt
# START OF FILE requirements.txt

# Bot framework for handling Telegram Bot API
python-telegram-bot==21.0.1

# Library for automating user accounts (Telethon client)
telethon==1.34.0

# For rich, colorful logging in the console
rich==13.7.1

# Required for the web server on Vercel
Flask==3.0.3

# Required by APScheduler's SQLiteJobStore (can be kept or removed, Telethon might use it)
SQLAlchemy==2.0.29